bug fix for deletion of first, in the list, acl entry

+++ arc-expenses-backend/service/src/main/java/arc/expenses/service/AclService.java
@@ -31,7 +31,7 @@ public class AclService extends JdbcMutableAclService{
             MutableAcl acl = (MutableAcl) readAclById(objectIdentity);
             for(int i=0; i<acl.getEntries().size();i++){
                 for(Sid sid : principals) {
-                    if(acl.getEntries().get(i).getSid().equals(sid) && !acl.getEntries().get(i).getPermission().equals(ArcPermission.READ)){
+                    if(acl.getEntries().get(i).getSid().equals(sid) && !acl.getEntries().get(i).getPermission().equals(ArcPermission.READ) && !acl.getEntries().get(i).getPermission().equals(ArcPermission.CANCEL)){
                         acl.deleteAce(i);
                         i = (i==1 ? 0:i--);
                     }
